pyproject.toml
[tool.poetry]
name = "empathy-engine"
version = "0.1.0"
description = "Emotion-aware TTS engine that modulates vocal characteristics based on detected emotions in text."
authors = ["EmpathyEngine Auto-Init <dev@example.com>"]
readme = "README.md"
packages = [{ include = "empathy_engine" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
transformers = "^4.46.0"
torch = "^2.4.0"
spacy = "^3.7.0"
faiss-cpu = "^1.9.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
numpy = "^2.1.0"
sentence-transformers = "^3.0.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
ruff = "^0.6.0"
black = "^24.8.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

